refactor(consolidate bolt logic to bolt_util)
